dlna: first cut at support for samsung subtitles
Not yet working for my F series, and I'm not sure why, but in theory
this implements external srt support for older Samsung devices.

Older Samsung devices use a combination of vendor extension XML and
additional request headers when retrieving the media file.  Only one
subtitle track is supported, and it must be an .srt with the same
basename as the video file.

While ugly, in order to ensure consistentcy in the output, a
BrowseMetadata request will now scan for associated resources
(aka subtitles.)  Relying on the vfs cache to make this less
painful than it appears.

Also providing timestamp info on CDS objects.